requirements.txt
absl-py==1.4.0
addict==2.4.0
adversarial-robustness-toolbox==1.13.0
aiohttp==3.8.4
aioice==0.7.6
aiortc==1.3.2
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autograd-lib==0.0.7
av==9.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessings==1.7
blinker==1.6.2
botorch==0.8.1
cachetools==5.2.1
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleverhans==4.0.0
click==8.1.3
cloudpickle==2.2.0
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
datasets==2.13.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dm-tree==0.1.8
dnspython==2.3.0
docker==6.0.1
dpcontracts==0.6.0
easydict==1.10
entrypoints==0.4
evaluate==0.4.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.1.2
Flask-JWT-Extended==4.4.3
Flask-SQLAlchemy==2.5.1
flatbuffers==23.5.26
fonttools==4.38.0
forbiddenfruit==0.1.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.5.3
gin-config==0.5.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
GPUtil==1.4.0
gpytorch==1.9.1
greenlet==2.0.2
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
huggingface-hub==0.16.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jstyleson==0.0.2
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaleido==0.2.1
kiwisolver==1.4.4
kornia==0.6.9
lightning-utilities==0.8.0
linear-operator==0.3.0
lit==15.0.7
llvmlite==0.39.1
loguru==0.6.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mnist==0.2.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.14
names==0.3.0
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.8.8
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
nvflare==2.3.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-gpu==1.15.1
onnxruntime-training==1.15.1
opencv-python==4.8.0.74
openvino==2023.0.1
openvino-dev==2023.0.1
openvino-telemetry==2023.1.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.12.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.1b0
pyee==9.0.4
Pygments==2.14.0
PyJWT==2.6.0
pylibsrtp==0.7.1
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.4
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning==2.0.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.1.3
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.31
sqlitedict==2.1.0
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.24.1
sympy==1.11.1
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-probability==0.19.0
terminado==0.17.1
testpath==0.6.0
texttable==1.6.7
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torch-ort==1.15.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
transformers==4.30.2
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
urllib3==1.26.14
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.14.1
xxhash==3.2.0
yacs==0.1.8
yarl==1.9.2
zipp==3.11.0
